+++ Fabric/src/main/java/fuzs/puzzleslib/fabric/impl/core/FabricModConstructor.java
@@ -19,5 +19,6 @@ public final class FabricModConstructor implements ModConstructorImpl<ModConstru
         modConstructor.onRegisterGameRegistries(new GameRegistriesContextFabricImpl());
         modConstructor.onRegisterDataPackRegistries(new DataPackRegistriesContextFabricImpl());
         modConstructor.onRegisterVillagerTrades(new VillagerTradesContextFabricImpl());
+        modConstructor.onAddDataPackReloadListeners(new DataPackReloadListenersContextFabricImpl());
     }
 }